Initial import
With this, things are mostly in place.  Any arbitrary pair of ranges
can be compared with magit-tbdiff-ranges, and magit-tbdiff-revs and
magit-tbdiff-revs-with-base provide analogs to the "A...B" and "BASE A
B" command-line shortcuts, respectively.

The magit-tbdiff-mode buffer seems to work well enough for inspecting
tbdiff's output, but there are a few issues.

  * When the parent section of a diff is active, the diff's
    highlighting is overridden by the magit-section-highlight.

  * Unlike regular diff hunks, interdiff hunks can't be applied.  We
    should probably intercept apply commands to raise a user-error
    instead of letting the code execute until we hit an uglier error.

  * It'd be nice to get magit-go-backward and magit-go-forward working
    in magit-tbdiff-mode buffers, but it's not straightforward because
    magit-xref-setup hardcodes the modes that it supports.