magit-tbdiff-wash: Use 'commit' section type
Replace custom section type with 'commit'.  The main advantage of this
is that the diff sections under an active parent are now highlighted
like other diffs in Magit instead of being shown in the
magit-section-highlight face.  (magit-diff-highlight provides special
handling of commit sections.)

As a bonus, a tbdiff assignment line now has a type and value that can
be used by commands like magit-visit-thing.  These lines consist of
either one or two commits.  When there are two commits, let's choose
the second one as the value of the section.  This is consistent with
tbdiff showing the subject line of the second commit and with changes
from the second commit being shown as additions, indicating that the
second range is considered the newer series.

These values don't seem that useful in the context of interdiffs, but
it's also not obvious what value (and section type) would be
better---and better to the point that it'd be worth coming up with
another solution to the highlighting issue above.