Substitute "interdiff" for "range diff", mostly
Don't use "interdiff" for the reason explained in the previous commit.
I'm going with "range diff" for now because, of the proposals that
were made on the in-progress patch series for the core analog of
tbdiff [1], I liked the "range-diff" suggestion [2] the best.

The only remaining occurrence is the description for
magit-tbdiff-popup's key in magit-diff-popup.  Don't change it there
since it's the reason that the key is 'i'.

[1] https://public-inbox.org/git/cover.1525361419.git.johannes.schindelin@gmx.de/
[2] https://public-inbox.org/git/CAGZ79kZbRCH2OiTW1Ge31R9JN+vWD6tcjNWVGSzkSBcYZvwDjw@mail.gmail.com/