Support range-diff's --dual-color
This uses an approach that follows what Magit uses to support
--color-moved, but there are two additional aspects:

  * Enabling colored output for --dual-color means that the assignment
    lines are also colored, which we don't want.  Remove the ANSI
    escape sequences from the assignment lines.

  * range-diff --dual-color shows the output diff markers as inverted.
    ansi-color doesn't handle this in a way that is useful for our
    purposes because it maps the inverted colors to the error face.

    Instead put a box around the inverted text.

Note that the dual color presentation for the outer diff hunks don't
make it through the wash, but retaining their dual color isn't useful
within Magit's interface because the hunk sections already serve to
distinguish the outer hunk markers.

Don't add --dual-color to the transient when tbdiff is the subcommand
because I've only tested this with range-diff.  The approach here
should work for tbdiff too, except that tbdiff doesn't offer a way to
set other diff parts to "normal".  If there are users that can't
upgrade to a git with range-diff but want --dual-color, we can revisit
exposing the flag for tbdiff as well.

Closes #3.