Use custom wash-sequence function for Git 2.23 compatibility
magit-tbdiff-wash fails on the range-diff output from Git 2.23,
specifically the changes introduced by 43ba21cb57 (Merge branch
'tg/range-diff-output-update', 2019-07-25).  That series includes a
few improvements, but the relevant change for our parsing is that the
outer hunk headings are now processed to look like this:

    @@ Information

This presentation, in contrast to raw hunk headings, is an improvement
for human readers (and makes sense given the purpose of range-diff).
But it means that the outer diffs no longer have valid headings, so we
can't feed them blindly to magit-diff-wash-hunk.  Add a custom
-wash-hunk variant to handle these.

Sadly hunk refinement no longer works with the new range-diff output
because diff-refine-hunk (reasonably) does not recognize the hunk
header.

Closes #5.