Avoid standard hunk painting with --dual-color
range-diff's output can be confusing due to the two layers of diff
output.  --dual-color can make things clearer, but its presentation
doesn't play well with standard hunk highlighting on dark backgrounds.
(See gh-6 for some images.)

Avoid treating --dual-color hunks as regular hunks by using a distinct
section, resulting in output that is much closer to range-diff's
output in the terminal.  This solution is a bit ugly in that the
tbdiff hunk sections have different types depending on whether
--dual-color is used.  To avoid that, the same treatment could be used
when --dual-color is not specified: force color, feed the output to
ansi-color-apply-on-region, and use the same custom section.  But then
the hunks would lose visual consistency with regular hunks, which
doesn't seem like a good trade.

Closes #6.