Allow dual color for inner hunk headers
To support --dual-color, the diff is inserted with ANSI escape
sequences, and those are converted to overlays with
ansi-color-apply-on-region.  A few color.diff slots are set to normal
to avoid decorating them.  The frag and hunk slots are included in
this list.

I suspect I did that to avoid the overlays from messing with the
_outer_ header, which is turned into a section.  However, the last
commit deals with that directly, so let Git paint the hunk headers.